requirements.txt
asgiref==3.5.0
Django==3.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2